Save and Update WishList

+++ src/main/java/com/example/ra/repository/WishListRepository.java
@@ -4,5 +4,5 @@ import com.example.ra.model.entity.Category;
 import com.example.ra.model.entity.WishList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WishListRepository extends JpaRepository<WishList,Long> {
+public interface WishListRepository extends JpaRepository<WishList,Integer> {
 }
